requirements.txt
hikari==2.0.0.dev122
python-dotenv==1.0.0